refactor

+++ src/ft_list/ft_lstsize.c
@@ -6,15 +6,15 @@
 /*   By: mleblanc <mleblanc@student.42quebec.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/05/07 10:47:57 by mleblanc          #+#    #+#             */
-/*   Updated: 2021/06/15 20:23:25 by mleblanc         ###   ########.fr       */
+/*   Updated: 2021/10/06 01:24:52 by mleblanc         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_lstsize(t_list *lst)
+size_t	ft_lstsize(t_list *lst)
 {
-	int	size;
+	size_t	size;
 
 	size = 0;
 	while (lst)
